fix: aggregation pattern
